+++ R/engine_breg.R
@@ -179,12 +179,16 @@ engine_breg <- function(x,
           if(length(any_missing)>0){
             presabs <- presabs[-any_missing,] # This works as they are in the same order
             model$biodiversity[[1]]$expect <- model$biodiversity[[1]]$expect[-any_missing]
+          }
+          df <- subset(df, stats::complete.cases(df))
+          assertthat::assert_that(nrow(presabs) == nrow(df))
+
+          # Check that expect matches
+          if(length(model$biodiversity[[1]]$expect)!=nrow(df)){
             # Fill the absences with 1 as multiplier. This works since absences follow the presences
             model$biodiversity[[1]]$expect <- c( model$biodiversity[[1]]$expect,
                                                  rep(1, nrow(presabs)-length(model$biodiversity[[1]]$expect) ))
           }
-          df <- subset(df, stats::complete.cases(df))
-          assertthat::assert_that(nrow(presabs) == nrow(df))
 
           # Overwrite observation data
           model$biodiversity[[1]]$observations <- presabs
@@ -321,7 +325,7 @@ engine_breg <- function(x,
         if(!is.null(li)) if(getOption('ibis.setupmessages')) myLog('[Estimation]','red',paste0("Package does not support custom link functions. Ignored!"))
         form <- model$biodiversity[[1]]$equation
         df <- cbind(model$biodiversity[[1]]$predictors,
-                    data.frame(observed = model$biodiversity[[1]]$observations[,'observed'])
+                    data.frame(observed = model$biodiversity[[1]]$observations[,'observed', drop = TRUE])
                     )
         df <- subset(df, select = c(model$biodiversity[[1]]$predictors_names, "observed"))
         w <- model$biodiversity[[1]]$expect # The expected exposure