+++ R/train.R
@@ -257,7 +257,8 @@ methods::setMethod(
       model[['predictors_names']] <- x$get_predictor_names()
       # Get predictor types
       lu <- sapply(model[['predictors']][model[['predictors_names']]], is.factor)
-      model[['predictors_types']] <- data.frame(predictors = names(lu), type = ifelse(lu,'factor', 'numeric') )
+      model[['predictors_types']] <- data.frame(predictors = names(lu), type = ifelse(lu,'factor', 'numeric'),
+                                                row.names = NULL)
       # Assign attribute to predictors to store the name of object
       model[['predictors_object']] <- x$predictors
       rm(lu)
@@ -462,6 +463,9 @@ methods::setMethod(
         } else { model[['biodiversity']][[id]][['pseudoabsence_settings']] <- getOption("ibis.pseudoabsence")}
       }
 
+      # convert observations to sf object first regardless of type
+      model$biodiversity[[id]]$observations <- guess_sf(model$biodiversity[[id]]$observations)
+
       # Aggregate observations if poipo
       if(aggregate_observations && model[['biodiversity']][[id]][['type']] == "poipo"){
         model$biodiversity[[id]]$observations <- aggregate_observations2grid(
@@ -474,9 +478,17 @@ methods::setMethod(
       }
 
       # Now extract coordinates and extract estimates, shifted to raster extraction by default to improve speed!
-      env <- get_rastervalue(coords = guess_sf(model$biodiversity[[id]]$observations),
+      env <- get_rastervalue(coords = model$biodiversity[[id]]$observations,
                              env = model$predictors_object$get_data(df = FALSE),
                              rm.na = FALSE)
+
+      # select only columns needed by equation
+      if (model$biodiversity[[id]]$equation != "<Default>") {
+
+        env <- subset(env, select = c("ID", "x", "y", attr(stats::terms.formula(model$biodiversity[[id]]$equation),
+                                                           "term.labels")))
+
+      }
 
       # Remove missing values as several engines can't deal with those easily
       miss <- stats::complete.cases(env)
@@ -556,8 +568,8 @@ methods::setMethod(
 
       # Save predictors extracted for biodiversity extraction
       model[['biodiversity']][[id]][['predictors']] <- env
-      model[['biodiversity']][[id]][['predictors_names']] <- model[['predictors_names']][which( model[['predictors_names']] %notin% co )]
-      model[['biodiversity']][[id]][['predictors_types']] <- model[['predictors_types']][model[['predictors_types']]$predictors %notin% co,]
+      model[['biodiversity']][[id]][['predictors_names']] <- names(env)[names(env) %notin% c("ID", "x", "y", "Intercept")]
+      model[['biodiversity']][[id]][['predictors_types']] <- model[['predictors_types']][model[['predictors_types']][, "predictors"] %in% names(env), ]
     }
 
     # If the method of integration is weights and there are more than 2 datasets, combine